requirements.txt
Flask
mysql-connector-python
speechrecognition
pyttsx3
werkzeug
